refact result button group
